Phase 4: architecture refactors — extract paths, break cycle, delete orphans
A1 - mcp-tools/types.ts path helpers extracted to utils/paths.ts:
- getProjectCwd / getMonomindDataRoot / migrateLegacyStoreFile moved
- mcp-tools/types.ts re-exports them (backward compat for existing imports)
- dependency direction is now correct: tool layer consumes path infra,
  not the reverse (every layer was importing from mcp-tools/ just to
  resolve the project root)

A2 - circular dep broken between mcp-client.ts and monomind-tools.ts:
- monomind-tools.ts:13 used to statically import searchNonCoreTools
  from mcp-client.ts; mcp-client.ts:62 lazy-imports monomind-tools.ts
- the static cycle worked at runtime because of the dynamic-import
  asymmetry, but a future refactor that switched the client side to a
  static import would close the loop and break
- monomind-tools now does a dynamic import() inside the handler
  (the cycle is broken on both directions, not just one)

A6 - deleted 4 orphan workspace packages that contained only stale
build artifacts (no source, no package.json):
- @monomind/graph (only node_modules)
- @monomind/security (only __tests__/coverage/)
- @monoes/monoplaybook (only dist/)
- plugins/agentic-qe (only node_modules)

Tests: 884 passed (unchanged). All 53 test files green.

Deferred (multi-day refactors out of session scope):
- A3 split init/executor.ts (3,067 lines)
- A4 split mcp-tools/monograph-tools.ts (2,460 lines)
- A5 split OrgDaemon (1,690 lines)
- A7 consolidate duplicated sanitizeError/validateString helpers
